pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "par"
version = "0.1.0"
description = "Prometheus Alert Rule Linter"
readme = "README.md"
requires-python = ">=3.9"
dependencies = ["pyyaml>=6.0"]

[project.scripts]
par = "par.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["par*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
